Move Resource to its own file
