+++ cfgov/scripts/_atomic_helpers.py
@@ -310,7 +310,7 @@ resource_list_show_thumbnails_false = {
     "type": "resource_list",
     "value": {
         "heading": "Test Resource List",
-        "snippet_type": "v1.models.snippets.Resource",
+        "snippet_type": "v1.models.resources.Resource",
         "show_thumbnails": False,
     }
 }
@@ -319,7 +319,7 @@ resource_list_show_thumbnails_true = {
     "type": "resource_list",
     "value": {
         "heading": "Test Resource List",
-        "snippet_type": "v1.models.snippets.Resource",
+        "snippet_type": "v1.models.resources.Resource",
         "show_thumbnails": True,
     }
 }
@@ -328,7 +328,7 @@ resource_list_actions_column_width_40 = {
     "type": "resource_list",
     "value": {
         "heading": "Test Resource List",
-        "snippet_type": "v1.models.snippets.Resource",
+        "snippet_type": "v1.models.resources.Resource",
         "actions_column_width": "40",
     }
 }
