+++ cfgov/v1/models/__init__.py
@@ -21,9 +21,8 @@ from v1.models.learn_page import (
 )
 from v1.models.menu_item import MenuItem
 from v1.models.newsroom_page import LegacyNewsroomPage, NewsroomPage
-from v1.models.snippets import (
-    Contact, Resource, ResourceTag, ReusableText, TaggableSnippetManager
-)
+from v1.models.resources import Resource, ResourceTag, TaggableSnippetManager
+from v1.models.snippets import Contact, ReusableText
 from v1.models.sublanding_filterable_page import (
     ActivityLogPage, SublandingFilterablePage
 )
